requirements.txt
clipboard~=0.0.4
pytesseract~=0.3.13
pillow~=11.0.0